Data Extraction/requirements.txt
pathlib==1.0.1
snowflake_connector_python==2.3.2
pandas==1.1.2
boto3==1.14.63
jsonschema==3.2.0
xmltodict==0.12.0
lxml==4.8.0
pysftp==0.2.9
pyodbc==4.0.30
cx-Oracle~=8.2.1
aws-lambda-powertools
connectorx